+++ isra/main.py
@@ -10,6 +10,7 @@ from isra.src.ile import ile
 from isra.src.screening import screening
 from isra.src.standards import standards
 from isra.src.tests import tests
+from isra.src.tm import tm
 
 app = typer.Typer(no_args_is_help=True, add_completion=False, add_help_option=False)
 app_completion = typer.Typer(help="Generate and install completion scripts.", hidden=True)
@@ -20,6 +21,7 @@ app.add_typer(ile.app, name="ile", hidden=True)
 app.add_typer(screening.app, name="screening")
 app.add_typer(standards.app, name="standards")
 app.add_typer(component.app, name="component")
+app.add_typer(tm.app, name="tm", hidden=True)
 
 config.initialize_configuration()
 
@@ -54,4 +56,4 @@ def about():
 
 if __name__ == "__main__":
     config.initialize_configuration()
-    app(["about"])
+    app(["component", "batch"])