+++ isra/src/screening/screening.py
@@ -2,12 +2,13 @@ import typer
 
 from isra.src.config.config import get_sf_values
 from isra.src.config.constants import IR_SF_T_STRIDE, IR_SF_C_SCOPE, IR_SF_C_STANDARD_BASELINES
-from isra.src.screening.screening_service import get_all_threats, screening, get_stride_category, save_stride_category, \
+from isra.src.screening.screening_service import get_all_threats, screening, get_stride_category, \
+    save_stride_category, \
     get_attack_technique, save_attack_technique, get_attack_mitigation, save_attack_mitigation, get_all_controls, \
     get_intended_scope, save_intended_scope, get_baseline_standard_ref, save_baseline_standard_ref, \
     get_baseline_standard_section, save_baseline_standard_section, get_cia_triad, save_cia_triad, get_proper_cost, \
-    save_proper_cost, get_proper_cwe, save_proper_cwe, generate_question, save_question, threat_generator, \
-    control_generator, autoscreening_init, fix_component
+    save_proper_cost, get_proper_cwe, save_proper_cwe, generate_question, save_question, \
+    autoscreening_init, fix_component
 
 app = typer.Typer(no_args_is_help=True, add_help_option=False)
 
@@ -115,22 +116,6 @@ def question():
 
     items = get_all_controls()
     screening(items, generate_question, save_question)
-
-
-@app.command()
-def new_threat():
-    """
-    Generates a threat based on the current component name and description
-    """
-    threat_generator()
-
-
-@app.command()
-def new_control():
-    """
-    Generates a countermeasure based on the current component name and description
-    """
-    control_generator()
 
 
 @app.command()